+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.24.0] - 2020-12-28
+### Changed
+- **BREAKING CHANGE** `read` and `write` methods now return
+  `Result<usize, FtStatus>` where `usize` is the the number of bytes read or
+  written.
+  Previous `read` and `write` functionality that returned
+  `Result<(), TimeoutError>` is replaced by `read_all` and `write_all`.
+
 ## [0.23.0] - 2020-10-23
 ### Changed
 - `synchronize_mpsse` will now timeout if no read data is received and a read
@@ -56,7 +64,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Prior releases
 A changelog was not kept for prior releases.
 
-[Unreleased]: https://github.com/newAM/libftd2xx-rs/compare/0.23.0...HEAD
+[Unreleased]: https://github.com/newAM/libftd2xx-rs/compare/0.24.0...HEAD
+[0.24.0]: https://github.com/newAM/libftd2xx-rs/releases/tag/0.24.0
 [0.23.0]: https://github.com/newAM/libftd2xx-rs/releases/tag/0.23.0
 [0.22.0]: https://github.com/newAM/libftd2xx-rs/releases/tag/0.22.0
 [0.21.1]: https://github.com/newAM/libftd2xx-rs/releases/tag/0.21.1